requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
csbdeep=0.7.4=pypi_0
jupyter=1.0.0=py311h06a4308_8
matplotlib=3.8.3=pypi_0
numba=0.59.0=pypi_0
numpy=1.26.4=pypi_0
pyyaml=6.0.1=py311h5eee18b_0
scikit-image=0.22.0=pypi_0
scipy=1.12.0=pypi_0
send2trash=1.8.2=py311h06a4308_0
stardist=0.8.5=pypi_0
